took Balances Page to a Balance component
